u-net/requirements.txt
numpy==2.2.6
pillow==11.0.0
torch==2.7.0+cu126
torchvision==0.22.0+cu126
tqdm==4.67.1